Add react-native-elements to project
