+++ View/TrackActivity/AddActivityInput/ActivityNameInput.js
@@ -1,6 +1,7 @@
 import React, {Component} from 'react';
 import PropTypes from 'prop-types';
 import {StyleSheet, Text, TextInput, View} from 'react-native';
+import {FormLabel, FormInput} from 'react-native-elements';
 
 
 class ActivityNameInput extends Component {
@@ -20,7 +21,7 @@ class ActivityNameInput extends Component {
     const {onNameChange, activityName} = this.props;
     return (
       <View style={styles.container}>
-        <TextInput
+        <FormInput
           onChange={onNameChange}
           value={activityName}
           keyboardType={'default'}
@@ -44,6 +45,7 @@ const styles = StyleSheet.create({
   },
 
   input: {
+    // This gives the input size, which allows us to interact with the element.
     flex: 1,
   },
 });
