+++ View/TrackActivity/AddActivityInput/ActivityTimeInput.js
@@ -1,30 +1,31 @@
 import React, {Component} from 'react';
 import PropTypes from 'prop-types';
 import {StyleSheet, Text, TextInput, View} from 'react-native';
+import {FormInput} from 'react-native-elements';
 
 
-class ActivityCategoryInput extends Component {
+class ActivityTimeInput extends Component {
 
   // --------------------------------------------------
   // Props
   // --------------------------------------------------
   static propTypes = {
-    onCategoryChange: PropTypes.func.isRequired,
-    activityCategory: PropTypes.string.isRequired,
+    onTimeChange: PropTypes.func.isRequired,
+    activityStartTime: PropTypes.string.isRequired,
   }
 
   // --------------------------------------------------
   // Render
   // --------------------------------------------------
   render() {
-    const {onCategoryChange, activityCategory} = this.props;
+    const {onTimeChange, activityStartTime} = this.props;
 
     return (
       <View style={styles.container}>
-        <TextInput
-          onChange={onCategoryChange}
-          value={activityCategory}
+        <FormInput
+          onChange={onTimeChange}
           keyboardType={'default'}
+          value={activityStartTime}
           style={styles.input}
         />
       </View>
@@ -46,9 +47,10 @@ const styles = StyleSheet.create({
   },
 
   input: {
+    // This gives the input size, which allows us to interact with the element.
     flex: 1,
   }
 });
 
 
-export default ActivityCategoryInput;
+export default ActivityTimeInput;
