+++ package.json
@@ -33,9 +33,10 @@
     "lodash": "^4.17.4",
     "merge-graphql-schemas": "^1.3.0",
     "morgan": "^1.9.0",
-    "native-base": "^2.3.3",
-    "pg": "^7.3.0",
+    "pg": "^7.4.0",
     "react": "16.0.0",
-    "react-native": "^0.48.4"
+    "react-native": "^0.50.0",
+    "react-native-elements": "^0.18.2",
+    "react-native-vector-icons": "^4.4.2"
   }
 }
